+++ src/Google.Analytics.SDK.Core/Hits/ItemHit.cs
@@ -1,26 +1,5 @@
-﻿using Google.Analytics.SDK.Core.Helper;
-
-namespace Google.Analytics.SDK.Core.Hits
+﻿namespace Google.Analytics.SDK.Core.Hits
 {
-    public abstract class ItemHitBase : HitBase
-    {
-        protected ItemHitBase(string transactionId)
-        {
-            HitType = HitTypes.Item;
-            TransactionId = transactionId;
-        }
-
-        public override bool IsValid()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public bool Validate()
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-
     public class ItemHit : ItemHitBase
     {
         public ItemHit(string transactionId) : base(transactionId)
