+++ src/Google.Analytics.SDK.Core/Hits/TimingHit.cs
@@ -1,24 +1,5 @@
-﻿using Google.Analytics.SDK.Core.Helper;
-
-namespace Google.Analytics.SDK.Core.Hits
+﻿namespace Google.Analytics.SDK.Core.Hits
 {
-    public abstract class TimingHitBase : HitBase
-    {
-        protected TimingHitBase()
-        {
-            HitType = HitTypes.Timing;
-        }
-        public override bool IsValid()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public bool Validate()
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-
     public class TimingHit : TimingHitBase
     {
         public TimingHit() : base()
