+++ src/Google.Analytics.SDK.Core/Hits/TimingHitBase.cs
@@ -1,0 +1,21 @@
+﻿using Google.Analytics.SDK.Core.Helper;
+
+namespace Google.Analytics.SDK.Core.Hits
+{
+    public abstract class TimingHitBase : HitBase
+    {
+        protected TimingHitBase()
+        {
+            HitType = HitTypes.Timing;
+        }
+        public override bool IsValid()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public bool Validate()
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
